+++ forge/src/main/java/dev/aika/smsn/forge/mixin/MixinInfo.java
@@ -18,6 +18,7 @@ public class MixinInfo {
                 "dev.aika.smsn.forge.mixin.projecte.ThreadCheckUUIDMixin", "projecte",
                 "dev.aika.smsn.forge.mixin.quark.ThreadContributorListLoaderMixin", "quark",
                 "dev.aika.smsn.forge.mixin.quark.QButtonMixin", "quark",
+                "dev.aika.smsn.forge.mixin.tenshilib.PatreonDataManagerMixin", "tenshilib",
                 "dev.aika.smsn.forge.mixin.titanium.RewardMixin", "titanium",
                 "dev.aika.smsn.forge.mixin.voidscape.DonatorHandlerMixin", "voidscape",
                 "dev.aika.smsn.forge.mixin.xaero.minimap.InternetMixin", "xaerominimap",